% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate.Vote}
\alias{animate.Vote}
\title{Animate a series of votes from a `Vote` object.}
\usage{
\method{animate}{Vote}(vote, movie.name, iter = NULL, ...)
}
\arguments{
\item{vote}{An object of class `Vote`.}

\item{movie.name}{Name of the GIF file.}

\item{iter}{A vector of the iterations to plot, defaults to `seq(1, vote$call$iter)`.}
}
\description{
`animate()` calls `animation::saveGIF()` to create a GIF file from plots of all `iter` created with \link{plot.Vote}.
}
