% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote-complex.R
\docType{methods}
\name{complex_voters}
\alias{complex_voters}
\alias{+,Voter,Voter-method}
\alias{+}
\alias{+,Voters,Voter-method}
\alias{+,Voter,Voters-method}
\alias{+,Voters,Voters-method}
\title{Aggregate Voters.}
\usage{
\S4method{+}{Voter,Voter}(e1, e2)

\S4method{+}{Voters,Voter}(e1, e2)

\S4method{+}{Voter,Voters}(e1, e2)

\S4method{+}{Voters,Voters}(e1, e2)
}
\arguments{
\item{e1, e2}{Objects of class \code{Voter} or \code{Voters}.}
}
\description{
The \code{+} function is used to combine multiple voters into the \code{Voters} class.
}
