% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_coalitions.R
\name{determine_coalitions}
\alias{determine_coalitions}
\title{Determine possible coalitions based on agenda setter, veto players and normal voters.}
\usage{
determine_coalitions(as, veto, normal)
}
\arguments{
\item{as}{Agenda Setter Index.}

\item{veto}{Veto Player Index.}

\item{normal}{Normal Voter Index.}
}
\description{
The agenda setter and veto player must always be part of the coalition.
This function returns possible combinations of AS+Veto+Normal that form a majority.
}
