% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote-classes.R
\docType{methods}
\name{initialize,SQ-method}
\alias{initialize,SQ-method}
\title{Constructor for SQ class.}
\usage{
\S4method{initialize}{SQ}(.Object, position, drift)
}
\description{
Constructor for SQ class.
}
