% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vote.R
\name{plot.Vote}
\alias{plot.Vote}
\title{Create a `ggplot2` plot detailing voter positions, status quo, voting outcome, voter indifference circles and winset.}
\usage{
\method{plot}{Vote}(vote, iter = 1, xlim = NULL, ylim = NULL,
  trace = NULL, ...)
}
\arguments{
\item{vote}{An object of class `Vote`.}

\item{iter}{An iteration for which the plot should be generated.}

\item{xlim}{Limits of the x-axis. Defaults to `NULL`.}

\item{ylim}{Limits of the y-axis. Defaults to `NULL`.}

\item{trace}{Integer indicating how many iterations back the status quo should be traced via a line segment. Defaults to `NULL`.}
}
\description{
Create a `ggplot2` plot detailing voter positions, status quo, voting outcome, voter indifference circles and winset.
}
