% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vote.R
\name{plot_vote}
\alias{plot_vote}
\alias{plot.Vote}
\alias{plot_roles}
\alias{plot_coalition}
\title{Visualize votes.}
\usage{
\method{plot}{Vote}(x, iter = 1, xlim = NULL, ylim = NULL,
  trace = NULL, ...)

plot_roles(x, iter = NULL)

plot_coalition(x, iter = NULL)
}
\arguments{
\item{x}{An object of class \code{Vote}.}

\item{iter}{The iteration for which the plot should be generated.}

\item{xlim}{Limits of the x-axis, defaults to \code{NULL}.}

\item{ylim}{Limits of the y-axis, defaults to \code{NULL}.}

\item{trace}{Integer indicating how many iterations back the status quo should be traced via a line segment, defaults to \code{NULL}.}

\item{...}{Additional keyword arguments.}
}
\description{
Plot one iteration of a \code{Vote} object, detailing voter positions, status quo, voting outcome, voter indifference circles and the winset.
}
