% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote-classes.R
\docType{class}
\name{Position-class}
\alias{Position-class}
\alias{Position}
\title{An S4 class representing a position, used to build Voter and Status Quo classes.}
\description{
An S4 class representing a position, used to build Voter and Status Quo classes.
}
\section{Slots}{

\describe{
\item{\code{position}}{A position vector.}

\item{\code{dimension}}{An integer specifying the dimension, defaults to `length(position)`}
}}

