% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote-classes.R
\docType{methods}
\name{initialize,Position-method}
\alias{initialize,Position-method}
\alias{is.position}
\title{Init function of the Position class.
Assigns the `position` vector and `dimension`, with default `dimension` being `length(position)`.}
\usage{
\S4method{initialize}{Position}(.Object, position)

is.position(x)
}
\arguments{
\item{position}{The position vector.}

\item{x}{An object.}
}
\description{
Init function of the Position class.
Assigns the `position` vector and `dimension`, with default `dimension` being `length(position)`.

Check whether an object is of class `Position`.
}
