% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting.R
\name{solve_vote}
\alias{solve_vote}
\title{Solve a voting problem using \code{ECOSolver}.}
\usage{
solve_vote(prob, dimension)
}
\arguments{
\item{prob}{A problem of class `CVXR::Problem`}

\item{dimension}{The dimension of the problem.}
}
\description{
Solve a voting problem using \code{ECOSolver}.
}
