% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote-classes.R
\docType{class}
\name{SQ-class}
\alias{SQ-class}
\alias{.SQ}
\alias{SQ}
\alias{SQ}
\title{S4 class representing the Status Quo.}
\usage{
SQ(position)
}
\arguments{
\item{position}{A position vector.}
}
\description{
S4 class representing the Status Quo.

Constructor for the SQ class.
}
\section{Slots}{

\describe{
\item{\code{position}}{A position vector.}

\item{\code{dimension}}{}
}}

