% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote-df.R
\name{vote_df}
\alias{vote_df}
\title{Convert a \code{Vote} object into a tidy dataframe.}
\usage{
vote_df(vote)
}
\arguments{
\item{vote}{A \code{Vote} object.}
}
\description{
Convert a \code{Vote} object into a tidy dataframe.
}
