% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote-classes.R
\docType{class}
\name{Voter}
\alias{Voter}
\alias{.Voter}
\alias{initialize,Voter-method}
\alias{Voter}
\alias{Veto}
\alias{AS}
\alias{RandomVoter}
\title{An S4 class representing a voter.}
\usage{
\S4method{initialize}{Voter}(.Object, position, role)

Voter(position, role = "Normal")

Veto(position)

AS(position)

RandomVoter(position)
}
\arguments{
\item{position}{A position vector.}

\item{role}{Defaults to `"Normal"`.}

\item{position}{A position vector.}

\item{position}{A position vector.}

\item{position}{A position vector.}
}
\description{
An S4 class representing a voter.

Constructor for \link{Voter-class}.

Constructor for Voter, by default with role `"Normal"`.

Constructor for Voter with role `"Veto"`.

Constructor for Voter with role `"AS"`.

Constructor for Voter with role `"Random"`.
}
\section{Slots}{

\describe{
\item{\code{position}}{A position vector.}

\item{\code{dimension}}{An integer specifying the dimension, defaults to `length(position)`}

\item{\code{role}}{The role of the voter, one of `"AS"`, `"Veto"`, `"Normal"` or `"Random"`.}
}}

