% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote-classes.R
\docType{class}
\name{voter_class}
\alias{voter_class}
\alias{.Voter}
\alias{Voter}
\alias{Veto}
\alias{AS}
\alias{RandomVoter}
\title{Voter Classes}
\usage{
Voter(position, role = "Normal")

Veto(position)

AS(position)

RandomVoter(position)
}
\arguments{
\item{position}{A position vector or matrix, either indicating the voter's position once (notwithstanding drift and vibration) or for all voting iterations.}

\item{role}{(optional) The role of the voter, one of "Normal", "Veto", "AS" and "Random". Can only be used with the \code{Voter()} constructor function.}
}
\description{
Voter classes are used to represent voters having various roles (Normal, Voter, Agenda Setter).
}
\section{Slots}{

\describe{
\item{\code{position}}{A position vector or matrix.}

\item{\code{dimension}}{An integer specifying the dimension, defaults to \code{length(position)} if \code{position} is of type \code{vector} or \code{ncol(position)} if \code{position} is of type \code{matrix}.}

\item{\code{role}}{The role of the voter, one of "AS", "Veto", "Normal" and "Random".}
}}

