% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voters.R
\docType{class}
\name{Voters-class}
\alias{Voters-class}
\alias{.Voters}
\title{An S4 class representing a group of voters.}
\description{
An S4 class representing a group of voters.
}
\section{Slots}{

\describe{
\item{\code{voter_count}}{Number of voters.}
}}

