% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voters-class.R
\docType{class}
\name{voters_class}
\alias{voters_class}
\alias{Voters-class}
\alias{.Voters}
\alias{initialize,Voters-method}
\alias{Voters}
\alias{as.voters}
\alias{as.voters.Voter}
\title{Voters Class.}
\usage{
\S4method{initialize}{Voters}(.Object, position, dimension, role,
  voter_count)

Voters(position, dimension, role)

as.voters(voter)

\method{as.voters}{Voter}(voter)
}
\arguments{
\item{.Object}{The Voters object.}

\item{position}{A position vector or matrix, either indicating the voter's position once (notwithstanding drift and vibration) or for all voting iterations.}

\item{dimension}{Dimensionality of the voters' positions.}

\item{role}{(optional) The role of the voter, one of "Normal", "Veto", "AS" and "Random". Can only be used with the \code{Voter()} constructor function.}

\item{voter_count}{The number of voters.}

\item{voter}{Voter object.}
}
\description{
The Voters class is used to represent multiple voters at the same time.
}
\section{Slots}{

\describe{
\item{\code{position}}{A position matrix with \code{dimension * voter_count} columns. There is either one row for fixed positions (notwithstanding drift and vibration), or one row per vote iteration.}

\item{\code{dimension}}{An integer specifying the dimension of each voter's position, defaults to \code{ncol(position) / voter_count}.}

\item{\code{role}}{The role of the voter, one of "AS", "Veto", "Normal" or "Random".}

\item{\code{voter_count}}{The number of voters.}
}}

