% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winset.R
\name{get_winset}
\alias{get_winset}
\title{Retrieve winset spatial polygons from `Vote` objects.}
\usage{
get_winset(vote, iter = 1, dimension = 2, quadsegs = 100)
}
\arguments{
\item{vote}{An object of class `Vote`.}

\item{iter}{A vector of iterations for which to retrieve the winsets.}

\item{dimension}{Defaults to `2`.}

\item{quadsegs}{Number of linear segments used to approximate voter indifference circles.}
}
\description{
\code{get_winset()} converts the positions of voters at the selected iterations into spatial polygons with the given voter radii, and calculates the intersection of all coalition members' indifference circles, which is called the winset.
}
