% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote-classes.R
\docType{class}
\name{Position}
\alias{Position}
\alias{initialize,Position-method}
\alias{is.position}
\title{Assigns the `position` vector and `dimension`, with default \code{dimension} of \code{length(position)}.}
\usage{
\S4method{initialize}{Position}(.Object, position)

is.position(x)
}
\arguments{
\item{.Object}{the Position object.}

\item{position}{The position vector.}

\item{x}{An object.}
}
\description{
Assigns the `position` vector and `dimension`, with default \code{dimension} of \code{length(position)}.

Check whether an object is of class `Position`.
}
\section{Slots}{

\describe{
\item{\code{position}}{A position vector.}

\item{\code{dimension}}{An integer specifying the dimension, defaults to `length(position)`}
}}

