% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting.R
\name{Vote}
\alias{Vote}
\title{Vote.}
\usage{
Vote(formula, drift = NULL, vibration = NULL, iter = 1,
  keep_winset_objects = TRUE, ...)
}
\arguments{
\item{formula}{A formula with SQ on LHS and Voters on RHS.}

\item{iter}{Number of runs.}
}
\description{
Executes the spatial model of voting, solving the `Formula`
constrained optimization problem.
}
