% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote-classes.R
\docType{methods}
\name{initialize,Voter-method}
\alias{initialize,Voter-method}
\title{Constructor for Voter class.}
\usage{
\S4method{initialize}{Voter}(.Object, position, role)
}
\arguments{
\item{.Object}{The Voter object.}

\item{position}{The position vector.}

\item{role}{Role character or vector.}
}
\description{
Constructor for Voter class.
}
